Move ActionType to actions.ts
