+++ optaweb-tsp-ui/src/store/tsp/actions.ts
@@ -15,7 +15,22 @@
  */
 
 import { Client, Frame } from 'webstomp-client';
-import { ActionType, ILatLng, ITSPRouteWithSegments } from './types';
+import { ILatLng, ITSPRouteWithSegments } from './types';
+
+// *************************************************************************************************
+// Action types
+// *************************************************************************************************
+
+export enum ActionType {
+  SOLUTION_UPDATES_DATA = 'SOLUTION_UPDATES_DATA',
+  DELETE_LOCATION = 'DELETE_LOCATION',
+  ADD_LOCATION = 'ADD_LOCATION',
+  ADD_DEMO_LOCATION = 'ADD_DEMO_LOCATION',
+  CLEAR_SOLUTION = 'CLEAR_SOLUTION',
+  WS_CONNECT = 'WS_CONNECT',
+  WS_CONNECT_SUCCESS = 'WS_CONNECT_SUCCESS',
+  WS_CONNECT_FAILURE = 'WS_CONNECT_FAILURE',
+}
 
 // *************************************************************************************************
 // Action interfaces
